requirements.txt
Flask
psycopg2-binary
python-dotenv
gunicorn
fpdf
